Refactoring

+++ UserManagerLib/Administration.php
@@ -4,7 +4,7 @@ require_once 'Database.php';
 require_once 'User.php';
 require_once "HelperFunctions.php";
 require_once "Registration.php";
-require_once "Queryable.php";
+require_once "RunsSQL.php";
 
 /**
  * Created by PhpStorm.
@@ -21,7 +21,7 @@ require_once "Queryable.php";
  * @author Benjamin Arnold <benji.arnold@gmail.com>
  * @author Peter L. Kim <peterlk.dev@gmail.com>
  */
-class Administration extends Queryable
+class Administration extends RunsSQL
 {
 
     public static function createUser($firstName, $middleName, $lastName, $email, $site, $password, $authType = "REGULAR")
@@ -36,8 +36,18 @@ class Administration extends Queryable
 
     public static function editUserType($email, $authType)
     {
-        $sql = "UPDATE users SET auth_type = " . parent::PREPARED_STATEMENT_1 . " WHERE email = " . parent::PREPARED_STATEMENT_2;
-        parent::runSQLWithTwoClauses($sql, $email, $authType, false);
+        if (HelperFunctions::isUserAuthorized(1)) {
+            $sql = "UPDATE users SET auth_type = " . parent::PREPARED_STATEMENT_1 . " WHERE email = " . parent::PREPARED_STATEMENT_2;
+            $execution = parent::runSQLWithTwoClauses($sql, $email, $authType, false);
+
+            if ($execution != false) {
+                return true;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 
     /**
@@ -45,9 +55,18 @@ class Administration extends Queryable
      */
     public static function editUserAddRegistrationTo($userId, $websiteId)
     {
-        $sql = "INSERT INTO user_site_xref (user_id, site_id) VALUES (" . parent::PREPARED_STATEMENT_1 . ", " . parent::PREPARED_STATEMENT_2 . ");";
-        $results = parent::runSQLWithTwoClauses($sql, $userId, $websiteId, true);
-        return $results;
+        if (HelperFunctions::isUserAuthorized(1)) {
+            $sql = "INSERT INTO user_site_xref (user_id, site_id) VALUES (" . parent::PREPARED_STATEMENT_1 . ", " . parent::PREPARED_STATEMENT_2 . ");";
+            $execution = parent::runSQLWithTwoClauses($sql, $userId, $websiteId, true);
+
+            if ($execution != false) {
+                return $execution;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 
     /**
@@ -57,9 +76,18 @@ class Administration extends Queryable
      */
     public static function editUserRemoveRegistrationFrom($userId, $websiteId)
     {
-        $sql = "DELETE FROM user_site_xref WHERE user_id = " . parent::PREPARED_STATEMENT_1 . " AND site_id = " . parent::PREPARED_STATEMENT_2;
-        $results = parent::runSQLWithTwoClauses($sql, $userId, $websiteId, true);
-        return $results;
+        if (HelperFunctions::isUserAuthorized(1)) {
+            $sql = "DELETE FROM user_site_xref WHERE user_id = " . parent::PREPARED_STATEMENT_1 . " AND site_id = " . parent::PREPARED_STATEMENT_2;
+            $execution = parent::runSQLWithTwoClauses($sql, $userId, $websiteId, true);
+
+            if ($execution != false) {
+                return $execution;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 
     /**
@@ -68,10 +96,19 @@ class Administration extends Queryable
      */
     public static function getUserWithID($userId)
     {
-        $sql = "SELECT lastname, firstname, middlename, email  FROM users WHERE " .
-            "user_id = " . parent::PREPARED_STATEMENT_1 . " ORDER BY lastname";
-        $results = parent::runSQLWithOneClause($sql, $userId, true);
-        return $results;
+        if (HelperFunctions::isUserAuthorized(1)) {
+            $sql = "SELECT lastname, firstname, middlename, email  FROM users WHERE " .
+                "user_id = " . parent::PREPARED_STATEMENT_1 . " ORDER BY lastname";
+            $execution = parent::runSQLWithOneClause($sql, $userId, true);
+
+            if ($execution != false) {
+                return $execution;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 
     /**
@@ -80,10 +117,19 @@ class Administration extends Queryable
      */
     public static function getUserWithEmail($userEmail)
     {
-        $sql = "SELECT lastname, firstname, middlename, user_id FROM users WHERE " .
-            "email = " . parent::PREPARED_STATEMENT_1 . " ORDER BY lastname";
-        $results = parent::runSQLWithOneClause($sql, $userEmail, true);
-        return $results;
+        if (HelperFunctions::isUserAuthorized(1)) {
+            $sql = "SELECT lastname, firstname, middlename, user_id FROM users WHERE " .
+                "email = " . parent::PREPARED_STATEMENT_1 . " ORDER BY lastname";
+            $execution = parent::runSQLWithOneClause($sql, $userEmail, true);
+
+            if ($execution != false) {
+                return $execution;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 
     /**
@@ -93,10 +139,19 @@ class Administration extends Queryable
      */
     public static function getUsersWithFirstAndLastName($userFirstName, $userLastName)
     {
-        $sql = "SELECT lastname, firstname, middlename, email, user_id FROM users WHERE " .
-            " firstname = " . parent::PREPARED_STATEMENT_1 . " AND lastname = " . parent::PREPARED_STATEMENT_2 . " ORDER BY lastname";
-        $results = parent::runSQLWithTwoClauses($sql, $userFirstName, $userLastName, true);
-        return $results;
+        if (HelperFunctions::isUserAuthorized(1)) {
+            $sql = "SELECT lastname, firstname, middlename, email, user_id FROM users WHERE " .
+                " firstname = " . parent::PREPARED_STATEMENT_1 . " AND lastname = " . parent::PREPARED_STATEMENT_2 . " ORDER BY lastname";
+            $execution = parent::runSQLWithTwoClauses($sql, $userFirstName, $userLastName, true);
+
+            if ($execution != false) {
+                return $execution;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 
     /**
@@ -105,10 +160,19 @@ class Administration extends Queryable
      */
     public static function getUsersByFirstNameStartingWith($searchLetter)
     {
-        $sql = "SELECT firstname, middlename, lastname, email, user_id FROM users WHERE " .
-            "firstname LIKE " . parent::PREPARED_STATEMENT_1 . "% ORDER BY firstname ASC";
-        $results = parent::runSQLWithOneClause($sql, $searchLetter, true);
-        return $results;
+        if (HelperFunctions::isUserAuthorized(1)) {
+            $sql = "SELECT firstname, middlename, lastname, email, user_id FROM users WHERE " .
+                "firstname LIKE " . parent::PREPARED_STATEMENT_1 . "% ORDER BY firstname ASC";
+            $execution = parent::runSQLWithOneClause($sql, $searchLetter, true);
+
+            if ($execution != false) {
+                return $execution;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 
     /**
@@ -117,9 +181,18 @@ class Administration extends Queryable
      */
     public static function getUsersByLastNameStartingWith($searchLetter)
     {
-        $sql = "SELECT lastname, firstname, middlename, email, user_id FROM users WHERE lastname LIKE " . parent::PREPARED_STATEMENT_1 . "% ORDER BY lastname ASC";
-        $results = parent::runSQLWithOneClause($sql, $searchLetter, true);
-        return $results;
+        if (HelperFunctions::isUserAuthorized(1)) {
+            $sql = "SELECT lastname, firstname, middlename, email, user_id FROM users WHERE lastname LIKE " . parent::PREPARED_STATEMENT_1 . "% ORDER BY lastname ASC";
+            $execution = parent::runSQLWithOneClause($sql, $searchLetter, true);
+
+            if ($execution != false) {
+                return $execution;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 
     /**
@@ -128,11 +201,20 @@ class Administration extends Queryable
      */
     public static function getUsersRegisteredToThisSite($currentWebsite)
     {
-        $sql = "SELECT users.lastname, users.firstname, users.middlename, users.email, users.user_id FROM " .
-            "users INNER JOIN users ON users.user_id = user_site_xref.user_id INNER JOIN sites ON " .
-            "sites.site_id = user_site_xref.site_id WHERE site.site_name = " . parent::PREPARED_STATEMENT_1 . " ORDER BY lastname";
-        $results = parent::runSQLWithOneClause($sql, $currentWebsite, true);
-        return $results;
+        if (HelperFunctions::isUserAuthorized(1)) {
+            $sql = "SELECT users.lastname, users.firstname, users.middlename, users.email, users.user_id FROM " .
+                "users INNER JOIN users ON users.user_id = user_site_xref.user_id INNER JOIN sites ON " .
+                "sites.site_id = user_site_xref.site_id WHERE site.site_name = " . parent::PREPARED_STATEMENT_1 . " ORDER BY lastname";
+            $execution = parent::runSQLWithOneClause($sql, $currentWebsite, true);
+
+            if ($execution != false) {
+                return $execution;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 
     /**
@@ -141,11 +223,21 @@ class Administration extends Queryable
      */
     public static function getUsersWithType($userType)
     {
-        $sql = "SELECT users.lastname, users.firstname, users.middlename, users.email, users.user_id FROM " .
-            "users INNER JOIN users ON users.user_id = user_auth_xref.user_id INNER JOIN auth ON " .
-            "auth.auth_type = user_auth_xref.auth_type WHERE auth.auth_type = " . parent::PREPARED_STATEMENT_1 . " ORDER BY lastname";
-        $results = parent::runSQLWithOneClause($sql, $userType, true);
-        return $results;
+        if (HelperFunctions::isUserAuthorized(1)) {
+            $sql = "SELECT users.lastname, users.firstname, users.middlename, users.email, users.user_id FROM " .
+                "users INNER JOIN users ON users.user_id = user_auth_xref.user_id INNER JOIN auth ON " .
+                "auth.auth_type = user_auth_xref.auth_type WHERE auth.auth_type = " . parent::PREPARED_STATEMENT_1 . " ORDER BY lastname";
+
+            $execution = parent::runSQLWithOneClause($sql, $userType, true);
+
+            if ($execution != false) {
+                return $execution;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 
     /**
@@ -153,9 +245,18 @@ class Administration extends Queryable
      */
     public static function getWebsites()
     {
-        $sql = "SELECT * FROM sites";
-        $results = parent::runSQLWithNoClause($sql, true);
-        return $results;
+        if (HelperFunctions::isUserAuthorized(1)) {
+            $sql = "SELECT * FROM sites";
+            $execution = parent::runSQLWithNoClause($sql, true);
+
+            if ($execution != false ) {
+                return $execution;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 
     /**
@@ -164,45 +265,85 @@ class Administration extends Queryable
      */
     public static function deleteUserById($userId)
     {
-        $sql = "DELETE FROM users WHERE user_id = " . parent::PREPARED_STATEMENT_1 . "";
-        parent::runSQLWithOneClause($sql, $userId, false);
+        if (HelperFunctions::isUserAuthorized(1)) {
+            $sql = "DELETE FROM users WHERE user_id = " . parent::PREPARED_STATEMENT_1 . "";
+            $execution = parent::runSQLWithOneClause($sql, $userId, false);
+
+            if ($execution != false) {
+                return true;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 
 
     public static function addNewAuthType($newAuthType, $rankAssociated)
     {
-        $isAuthTypeUnique = !HelperFunctions::isValidType($newAuthType);
-        $isCharsAllUpperCase =  ctype_upper($newAuthType);
-        $isRankAnInt = ctype_digit($rankAssociated);
+        if (HelperFunctions::isUserAuthorized(1)) {
+            $isAuthTypeUnique = !HelperFunctions::isValidAuthType($newAuthType);
+            $isCharsAllUpperCase = ctype_upper($newAuthType);
+            $isRankAnInt = ctype_digit($rankAssociated);
 
-        if($isAuthTypeUnique && $isCharsAllUpperCase && $isRankAnInt && HelperFunctions::isUserAuthorized(0)) {
+            if ($isAuthTypeUnique && $isCharsAllUpperCase && $isRankAnInt && HelperFunctions::isUserAuthorized(0) && $rankAssociated > 1 && !HelperFunctions::isAuthRankValueAlreadyInUse($rankAssociated)) {
 
-            $sql = "INSERT INTO auth (auth_type, auth_rank) values( " . parent::PREPARED_STATEMENT_1 . ", " . parent::PREPARED_STATEMENT_2 . ");";
-            parent::runSQLWithTwoClauses($sql, $newAuthType, $rankAssociated, false);
+                $sql = "INSERT INTO auth (auth_type, auth_rank) values( " . parent::PREPARED_STATEMENT_1 . ", " . parent::PREPARED_STATEMENT_2 . ");";
+
+                $execution = parent::runSQLWithTwoClauses($sql, $newAuthType, $rankAssociated, false);
+
+                if ($execution != false) {
+                    return true;
+                } else {
+                    return false;
+                }
+
+            }
+
         }
-
     }
 
-    public static function deleteAuthType($authType){
+    public static function deleteAuthType($authType)
+    {
         if (HelperFunctions::isUserAuthorized(0)) {
-            if($authType != "SUPER_ADMIN" || $authType != "ADMIN") {
+            if ($authType != "SUPER_ADMIN" || $authType != "ADMIN") {
 
                 $sql = "DELETE FROM auth_user_xref WHERE auth_type = " . parent::PREPARED_STATEMENT_1;
-                parent::runSQLWithOneClause($sql, $authType, false);
+                $execution1 = parent::runSQLWithOneClause($sql, $authType, false);
 
                 $sql = "DELETE FROM auth WHERE auth_type = " . parent::PREPARED_STATEMENT_1;
-                parent::runSQLWithOneClause($sql, $authType, false);
+                $execution2 = parent::runSQLWithOneClause($sql, $authType, false);
+
+                if ($execution1 != false && $execution2 != false) {
+                    return true;
+                } else {
+                    return false;
+                }
+            } else {
+                return false;
             }
         }
     }
 
-    public static function deleteSite($siteName) {
+    public static function deleteSite($siteName)
+    {
+        if (HelperFunctions::isUserAuthorized(1)) {
 
-        $sql = "DELETE FROM user_site_xref WHERE site_name = " . parent::PREPARED_STATEMENT_1;
-        parent::runSQLWithOneClause($sql, $siteName, false);
+            $sql = "DELETE FROM user_site_xref WHERE site_name = " . parent::PREPARED_STATEMENT_1;
+            $execution1 = parent::runSQLWithOneClause($sql, $siteName, false);
 
-        $sql = "DELETE FROM sites WHERE site_name = " . parent::PREPARED_STATEMENT_1;
-        parent::runSQLWithOneClause($sql, $siteName, false);
+            $sql = "DELETE FROM sites WHERE site_name = " . parent::PREPARED_STATEMENT_1;
+            $execution1 = parent::runSQLWithOneClause($sql, $siteName, false);
+
+            if ($execution1 != false && $execution2 != false) {
+                return true;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 
 }
